requirements.txt
numpy==1.23.5
opencv-python==4.6.0.66
packaging==22.0
pdf2image==1.16.0
Pillow==9.3.0
pytesseract==0.3.10
